добавлен вкт

+++ transform.py
@@ -1,9 +1,12 @@
 from NGSatSearch.NGSatSearch import NGSatSearch
 import os
+import shutil
 import zipfile
-import json
+import geojson
+from shapely.geometry import shape
 from osgeo import gdal, ogr, osr
 import pyproj
+import json
 
 import numpy as np
 
@@ -14,14 +17,23 @@ def extract(folder):
     dir_list = os.listdir(folder)
     for directory in dir_list:
         dir_path = os.path.join(folder, directory)
-        with zipfile.ZipFile(dir_path, 'r') as zip_open:
-            zip_open.extractall(folder)
-        os.remove(dir_path)
-    print('Распаковка успешна')
+        if '.zip' in dir_path:
+            try:
+                with zipfile.ZipFile(dir_path, 'r') as zip_open:
+                    zip_open.extractall(folder)
+                    os.remove(dir_path)
+            except:
+                os.remove(dir_path)
+    print('Распаковка успешна')  # todo : добавить обработку плохих зипов (ниже ошибка)
+
+
+'''
+raise BadZipFile("File is not a zip file")
+zipfile.BadZipFile: File is not a zip file
+'''
 
 
 def transform_tiff(folder, boundary, polarization_type=None):
-
     if not os.path.isdir('transformed_image'):
         os.mkdir('transformed_image')
 
@@ -46,39 +58,64 @@ def transform_tiff(folder, boundary, polarization_type=None):
         if not os.path.isdir(f"transformed_image\\{tiff_file_name.upper()}"):
             os.mkdir(f"transformed_image\\{tiff_file_name.upper()}")
 
-        output_file = os.path.join('transformed_image', f'{tiff_file_name.upper()}', f'{tiff_file_name}_reproj.tiff')
+        output_file_dir = os.path.join('transformed_image', f'{tiff_file_name.upper()}')
+        tiff_file_basename = os.path.basename(tiff_file)
+        output_file = os.path.join(output_file_dir, tiff_file_basename)
+        shutil.copy(tiff_file, output_file)
 
         print('Файлы выбраны')
 
-        source_dataset = gdal.Open(tiff_file)
-        target_projection = 'EPSG:3857'
+        # src_dataset = gdal.Open(tiff_file)
+        # target_projection = 'EPSG:3857'
+        #
+        # if src_dataset is None:
+        #     raise Exception("Не удалось открыть исходный GeoTIFF.")
+        #
+        # # src_srs = src_dataset.GetGCPProjection()
+        # gcps = src_dataset.GetGCPs()
+        #
+        # src_width = src_dataset.RasterXSize
+        # src_height = src_dataset.RasterYSize
+        #
+        # translate_options = gdal.TranslateOptions(outputSRS=target_projection,
+        #                                           GCPs=gcps,
+        #                                           width=src_width,
+        #                                           height=src_height,
+        #                                           )
+        #
+        # gdal.Translate(output_file,
+        #                src_dataset,
+        #                options=translate_options)
 
-        if source_dataset is None:
-            raise Exception("Не удалось открыть исходный GeoTIFF.")
-
-        src_srs = source_dataset.GetGCPProjection()
-        gcps = source_dataset.GetGCPs()
-
-        translate_options = gdal.TranslateOptions(outputSRS=target_projection, GCPs=gcps)
-
-        gdal.Translate(output_file, source_dataset, options=translate_options)
+        # src_width = src_dataset.RasterXSize
+        # src_height = src_dataset.RasterYSize
+        #
+        # warp_options = gdal.WarpOptions(dstSRS=target_projection,
+        #                                 format='GTiff',
+        #                                 srcSRS='EPSG:4326',
+        #                                 width=src_width,
+        #                                 height=src_height
+        #                                 )
+        #
+        # gdal.Warp(output_file,
+        #           src_dataset,
+        #           options=warp_options)
 
         # tiff_files.append(output_file)
-
-        source_dataset = None
-
-        print('Перепроецирование завершено')
-
-        # boundary_rep = reproject_geojson(boundary)
         #
-        # cut_tiff = crop_tiff(output_file, boundary_rep, f"transformed_image\\{tiff_file_name.upper()}")
+        # source_dataset = None
+        #
+        # print('Перепроецирование завершено')
 
         channel_stat = calculating_percentiles(output_file)
         print(channel_stat)
 
-        # os.remove(cut_tiff)
-
         qml_generator(channel_stat, f"transformed_image\\{tiff_file_name.upper()}")
+
+        crop_tiff(output_file, boundary)
+        print('Обрезка завершена')
+
+        os.remove(output_file)
 
     # return output_file
 
@@ -90,17 +127,17 @@ def calculating_percentiles(tiff_file):
 
     channel_stat = {}
 
-    for num_channel in range(1, num_channels + 1):
-        channel = dataset.GetRasterBand(num_channel)
-        channel_data = channel.ReadAsArray()
-        per_5 = np.percentile(channel_data, 5)
-        per_95 = np.percentile(channel_data, 95)
-        # channel_stat.append((os.path.basename(tiff_file), per_5, per_95))
-        channel_stat = {
-            'tiff_file': f"{os.path.basename(tiff_file).split('.')[0]}",
-            'per_5': str(int(per_5)),
-            'per_95': str(int(per_95)),
-        }
+    # for num_channel in range(1, num_channels + 1):
+    channel = dataset.GetRasterBand(1)
+    channel_data = channel.ReadAsArray()
+    per_5 = np.percentile(channel_data, 5)
+    per_95 = np.percentile(channel_data, 95)
+    # channel_stat.append((os.path.basename(tiff_file), per_5, per_95))
+    channel_stat = {
+        'tiff_file': f"{os.path.basename(tiff_file).split('.')[0]}",
+        'per_5': str(int(per_5)),
+        'per_95': str(int(per_95)),
+    }
 
     dataset = None
 
@@ -124,8 +161,8 @@ def qml_generator(channel_stat, folder):
     return qml_tree.write(f"{folder}\\{channel_stat['tiff_file']}.qml", encoding='UTF-8')
 
 
-def crop_tiff(input_file, boundary, folder):
-    cut_tiff = os.path.join(folder, 'raster.tiff')
+def crop_tiff(input_file, boundary):
+    cut_tiff = input_file.replace('.tiff', '_cut.tiff')
 
     # gdal.SetConfigOption('CHECK_DISK_FREE_SPACE', 'FALSE')
 
@@ -146,8 +183,6 @@ def crop_tiff(input_file, boundary, folder):
     )
 
     del res  # Вызываем деструктор, чтобы записать данные на диск
-
-    return cut_tiff
 
 
 def reproject_geojson(input_file):
@@ -171,5 +206,26 @@ def reproject_geojson(input_file):
     return output_file
 
 
-# extract('images')
-# transform_tiff('images', 'boundary.geojson')
+extract('images')
+# transform_tiff('images - Copy', 'boundary.geojson')
+
+
+# def geojson_to_wkt(boundary):
+#     with open(boundary, 'r') as f:
+#         data = geojson.load(f)
+#
+#     wkt_geometries = []
+#
+#     if data['type'] == 'FeatureCollection':
+#         for feature in data['features']:
+#             geometry = feature['geometry']
+#             wkt_geometries.append(shape(geometry).wkt)
+#     else:
+#         geometry = data['geometry']
+#         wkt_geometries.append(shape(geometry).wkt)
+#
+#     return '\n'.join(wkt_geometries)
+#
+# wkt_geometries = geojson_to_wkt('boundary.geojson')
+#
+# print(wkt_geometries)
